requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
aiofiles==24.1.0
pillow==12.2.0
fastapi-mail==1.4.2
python-dotenv==1.0.1
pymysql==1.1.1
cryptography==44.0.2
requests
phonepe-pg-sdk-python
google-auth==2.38.0
fpdf2==2.8.3
slowapi==0.1.9
bleach==6.2.0
sentry-sdk[fastapi]==2.18.0
alembic==1.14.0
pytest==8.3.4
httpx==0.28.1
